Move TokenOrRequest to token.rs
TokenOrRequest doesn't seem GCP / service account specific, so in
prepartion for adding a TokenProvider trait, I figured I'd move
this into token.rs directly.

cargo test still passes.
