+++ src/token.rs
@@ -43,6 +43,32 @@ impl Token {
     }
 }
 
+#[derive(Debug)]
+pub enum RequestReason {
+    /// An existing token has expired
+    Expired,
+    /// The requested scopes have never been seen before
+    ScopesChanged,
+}
+
+/// Either a valid token, or an HTTP request that
+/// can be used to acquire one
+#[derive(Debug)]
+pub enum TokenOrRequest {
+    /// A valid token that can be supplied in an API request
+    Token(Token),
+    Request {
+        /// The parts of an HTTP request that must be sent
+        /// to acquire the token, in the client of your choice
+        request: http::Request<Vec<u8>>,
+        /// The reason we need to retrieve a new token
+        reason: RequestReason,
+        /// An opaque hash of the scope(s) for which the request
+        /// was constructed
+        scope_hash: u64,
+    },
+}
+
 impl std::convert::TryInto<http::header::HeaderValue> for Token {
     type Error = crate::Error;
